Everything works in this.
It prints individual reports. It prints a master report. The master report doesn't look like much. I'm not sure what to make it into. There are two templates, Master Template and Individual template. 
We may need a functionality to ask for the name of the qualtrix output file. Right now it is set to SP16 TA LA Eval.csv.